Change defgroup tag to not clobber ox-html
As described in https://github.com/marsmining/ox-twbs/issues/3, the current defgroup :tag for ox-twbs is the same as ox-html. This could lead to confusion among users who use emac's customize-group settings. This commit changes the :tag and the docstribg to better differentiate it from ox-html in the customize-group menu.